Implementing Product controller
